Split CUDA copy logic into per-side memcpy files
