+++ cuda_client.cpp
@@ -51,13 +51,13 @@
 #include "client_routing.h"
 #include "codegen/gen_cuda_client.h"
 #include "codegen/gen_rpc_ids.h"
+#include "cuda_client_memcpy.h"
 #include "cuda_profiler_compat.h"
 #include "events.h"
 #include "ipc.h"
 #include "lupine_attr_sizes.h"
 #include "lupine_fatbin.h"
 #include "lupine_log.h"
-#include "memcpy.h"
 #include "rpc.h"
 #include "third_party/libcuckoo/libcuckoo/cuckoohash_map.hh"
 #include "transport.h"
@@ -3153,54 +3153,15 @@ CUresult cuMemcpyHtoD_v2(CUdeviceptr dstDevice, const void *srcHost,
                          size_t ByteCount);
 CUresult cuMemcpyDtoH_v2(void *dstHost, CUdeviceptr srcDevice,
                          size_t ByteCount);
+
 CUresult cuMemcpyHtoDAsync_v2(CUdeviceptr dstDevice, const void *srcHost,
                               size_t ByteCount, CUstream hStream);
 CUresult cuStreamSynchronize(CUstream hStream);
-extern "C" CUresult lupine_copy_dtoh_on_stream(void *dstHost,
-                                               CUdeviceptr srcDevice,
-                                               size_t ByteCount,
-                                               CUstream hStream);
 // Defined with the capture wrappers below; nonzero only while this client holds
 // a stream capture open, where a pageable copy must stay a graph node.
 extern std::atomic<int> lupine_active_stream_captures;
 extern "C" CUresult cuStreamQuery_ptsz(CUstream hStream);
 extern "C" CUresult cuStreamSynchronize_ptsz(CUstream hStream);
-
-extern "C" CUresult
-lupine_cuMemcpyDtoD_via_client(CUdeviceptr dstDevice, CUdeviceptr srcDevice,
-                               size_t ByteCount, CUstream hStream, bool async) {
-  LUPINE_TRACE_LOG("LUPINE cross-route D2D via client dst="
-                   << reinterpret_cast<void *>(dstDevice)
-                   << " src=" << reinterpret_cast<void *>(srcDevice)
-                   << " bytes=" << ByteCount << " async=" << async
-                   << " stream=" << hStream);
-  if (ByteCount == 0) {
-    return CUDA_SUCCESS;
-  }
-
-  constexpr size_t chunk_size = 16 * 1024 * 1024;
-  std::vector<unsigned char> staging(std::min(ByteCount, chunk_size));
-  size_t offset = 0;
-  while (offset < ByteCount) {
-    size_t chunk = std::min(staging.size(), ByteCount - offset);
-    CUresult result =
-        cuMemcpyDtoH_v2(staging.data(), srcDevice + offset, chunk);
-    if (result != CUDA_SUCCESS) {
-      return result;
-    }
-    if (async) {
-      result = cuMemcpyHtoDAsync_v2(dstDevice + offset, staging.data(), chunk,
-                                    hStream);
-    } else {
-      result = cuMemcpyHtoD_v2(dstDevice + offset, staging.data(), chunk);
-    }
-    if (result != CUDA_SUCCESS) {
-      return result;
-    }
-    offset += chunk;
-  }
-  return CUDA_SUCCESS;
-}
 
 #ifdef cuStreamDestroy
 #undef cuStreamDestroy
@@ -5327,495 +5288,6 @@ lupine_cuOccupancyMaxActiveBlocksPerMultiprocessorWithFlags_safe(
       numBlocks, func, blockSize, dynamicSMemSize, flags);
 }
 
-extern "C" CUresult cuMemcpyAtoH_v2(void *dstHost, CUarray srcArray,
-                                    size_t srcOffset, size_t ByteCount) {
-  lupine_route route = lupine_route_for_default();
-  if (lupine_route_is_local(route)) {
-    return lupine_call_real_cuda_fn("cuMemcpyAtoH_v2", dstHost, srcArray,
-                                    srcOffset, ByteCount);
-  }
-  CUresult return_value = CUDA_ERROR_DEVICE_UNAVAILABLE;
-  conn_t *conn = lupine_route_remote_conn(route);
-  if (lupine_prepare_rpc(conn) < 0 ||
-      rpc_write_start_request(conn, RPC_cuMemcpyAtoH_v2) < 0 ||
-      rpc_write(conn, &srcArray, sizeof(srcArray)) < 0 ||
-      rpc_write(conn, &srcOffset, sizeof(srcOffset)) < 0 ||
-      rpc_write(conn, &ByteCount, sizeof(ByteCount)) < 0) {
-    return CUDA_ERROR_DEVICE_UNAVAILABLE;
-  }
-  int request_id = rpc_write_end(conn);
-  if (request_id < 0 || rpc_read_start(conn, request_id) < 0) {
-    return CUDA_ERROR_DEVICE_UNAVAILABLE;
-  }
-
-  auto *copy_dst = static_cast<unsigned char *>(dstHost);
-  size_t offset = 0;
-  do {
-    size_t chunk =
-        std::min(ByteCount - offset, (size_t)LUPINE_COMPRESS_BLOCK_BYTES);
-    if (rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        (return_value == CUDA_SUCCESS && chunk != 0 &&
-         rpc_read(conn, copy_dst + offset, chunk) < 0)) {
-      rpc_read_end(conn);
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    bool final_chunk =
-        return_value != CUDA_SUCCESS || offset + chunk == ByteCount;
-    if (rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    if (return_value != CUDA_SUCCESS) {
-      return return_value;
-    }
-    offset += chunk;
-    if (!final_chunk && rpc_read_start(conn, request_id) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-  } while (offset < ByteCount);
-  return return_value;
-}
-
-#ifdef cuMemcpyAtoH
-#undef cuMemcpyAtoH
-#endif
-extern "C" CUresult cuMemcpyAtoH(void *dstHost, CUarray srcArray,
-                                 size_t srcOffset, size_t ByteCount) {
-  return cuMemcpyAtoH_v2(dstHost, srcArray, srcOffset, ByteCount);
-}
-
-extern "C" CUresult cuMemcpyDtoHAsync_v2(void *dstHost, CUdeviceptr srcDevice,
-                                         size_t ByteCount, CUstream hStream) {
-  // "API synchronization behavior" lets the driver make this copy synchronous
-  // when the destination is pageable, and every driver does. Callers rely on it
-  // instead of synchronizing: cuSPARSE reads its scalar results (nnz counts,
-  // SpVV dot products) that way, and a deferred copy either never lands or
-  // lands on a stack frame the caller has already left. The bytes come back
-  // through the same bounded, chunked path the synchronous copy uses, issued on
-  // this stream so they stay ordered behind its prior work. Capture is the
-  // exception: there the copy only becomes a graph node and must not block.
-  if (ByteCount != 0 && !lupine_host_ptr_is_page_locked(dstHost) &&
-      lupine_active_stream_captures.load(std::memory_order_relaxed) == 0) {
-    return lupine_copy_dtoh_on_stream(dstHost, srcDevice, ByteCount, hStream);
-  }
-
-  conn_t *conn = lupine_rpc_conn_for_deviceptr(srcDevice);
-  if (lupine_prepare_rpc(conn) < 0 ||
-      rpc_write_start_request(conn, RPC_cuMemcpyDtoHAsync_v2) < 0 ||
-      rpc_write(conn, &dstHost, sizeof(dstHost)) < 0 ||
-      rpc_write(conn, &srcDevice, sizeof(srcDevice)) < 0 ||
-      rpc_write(conn, &ByteCount, sizeof(ByteCount)) < 0 ||
-      rpc_write(conn, &hStream, sizeof(hStream)) < 0 ||
-      rpc_write_end(conn) < 0) {
-    return CUDA_ERROR_DEVICE_UNAVAILABLE;
-  }
-  if (ByteCount != 0) {
-    lupine_event_stale_semaphore_add(1);
-  }
-  return CUDA_SUCCESS;
-}
-
-#ifdef cuMemcpyDtoHAsync
-#undef cuMemcpyDtoHAsync
-#endif
-extern "C" CUresult cuMemcpyDtoHAsync(void *dstHost, CUdeviceptr srcDevice,
-                                      size_t ByteCount, CUstream hStream) {
-  return cuMemcpyDtoHAsync_v2(dstHost, srcDevice, ByteCount, hStream);
-}
-
-extern "C" CUresult cuMemcpy2D_v2(const CUDA_MEMCPY2D *pCopy) {
-  if (pCopy == nullptr) {
-    return CUDA_ERROR_INVALID_VALUE;
-  }
-  CUDA_MEMCPY2D copy = *pCopy;
-  // UNIFIED addresses are client-space: a managed alias is already the
-  // server's address for that allocation, a plain host pointer's bytes must
-  // travel, and a device pointer is already server-valid.
-  if (copy.srcMemoryType == CU_MEMORYTYPE_UNIFIED) {
-    if (!lupine_is_managed_host_alias(copy.srcDevice) &&
-        lupine_copy_pointer_is_host(copy.srcDevice)) {
-      copy.srcHost = (const void *)copy.srcDevice;
-      copy.srcMemoryType = CU_MEMORYTYPE_HOST;
-    }
-  }
-  if (copy.dstMemoryType == CU_MEMORYTYPE_UNIFIED) {
-    if (!lupine_is_managed_host_alias(copy.dstDevice) &&
-        lupine_copy_pointer_is_host(copy.dstDevice)) {
-      copy.dstHost = (void *)copy.dstDevice;
-      copy.dstMemoryType = CU_MEMORYTYPE_HOST;
-    }
-  }
-  bool src_host = copy.srcMemoryType == CU_MEMORYTYPE_HOST;
-  bool dst_host = copy.dstMemoryType == CU_MEMORYTYPE_HOST;
-  lupine_copy_direction direction =
-      src_host ? (dst_host ? lupine_copy_direction::host_to_host
-                           : lupine_copy_direction::host_to_device)
-               : (dst_host ? lupine_copy_direction::device_to_host
-                           : lupine_copy_direction::device_to_device);
-  const char *src_base =
-      (const char *)copy.srcHost + copy.srcY * copy.srcPitch + copy.srcXInBytes;
-  char *dst_base =
-      (char *)copy.dstHost + copy.dstY * copy.dstPitch + copy.dstXInBytes;
-  CUresult return_value;
-  switch (direction) {
-  case lupine_copy_direction::host_to_host:
-    for (size_t row = 0; row < copy.Height; ++row) {
-      memmove(dst_base + row * copy.dstPitch, src_base + row * copy.srcPitch,
-              copy.WidthInBytes);
-    }
-    return CUDA_SUCCESS;
-  case lupine_copy_direction::host_to_device: {
-    conn_t *conn = copy.dstMemoryType == CU_MEMORYTYPE_ARRAY
-                       ? lupine_rpc_conn_for_current_context()
-                       : lupine_rpc_conn_for_deviceptr(copy.dstDevice);
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy2D_v2) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_write_pitched(conn, src_base, copy.WidthInBytes, copy.Height,
-                          copy.srcPitch, 1, 0) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    if (return_value == CUDA_SUCCESS) {
-      return_value = lupine_sync_mapped_device_to_host();
-    }
-    return return_value;
-  }
-  case lupine_copy_direction::device_to_host: {
-    conn_t *conn = copy.srcMemoryType == CU_MEMORYTYPE_ARRAY
-                       ? lupine_rpc_conn_for_current_context()
-                       : lupine_rpc_conn_for_deviceptr(copy.srcDevice);
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy2D_v2) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        (return_value == CUDA_SUCCESS &&
-         rpc_read_pitched(conn, dst_base, copy.WidthInBytes, copy.Height,
-                          copy.dstPitch, 1, 0) < 0) ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    return return_value;
-  }
-  case lupine_copy_direction::device_to_device: {
-    if (copy.srcMemoryType != CU_MEMORYTYPE_ARRAY &&
-        copy.dstMemoryType != CU_MEMORYTYPE_ARRAY &&
-        !lupine_deviceptrs_share_route(copy.dstDevice, copy.srcDevice)) {
-      // Different devices: the rows route through the client.
-      for (size_t row = 0; row < copy.Height; ++row) {
-        return_value = lupine_cuMemcpyDtoD_via_client(
-            copy.dstDevice + (copy.dstY + row) * copy.dstPitch +
-                copy.dstXInBytes,
-            copy.srcDevice + (copy.srcY + row) * copy.srcPitch +
-                copy.srcXInBytes,
-            copy.WidthInBytes, nullptr, false);
-        if (return_value != CUDA_SUCCESS) {
-          return return_value;
-        }
-      }
-      return lupine_sync_mapped_device_to_host();
-    }
-    conn_t *conn;
-    if (copy.dstMemoryType != CU_MEMORYTYPE_ARRAY) {
-      conn = lupine_rpc_conn_for_deviceptr(copy.dstDevice);
-    } else if (copy.srcMemoryType != CU_MEMORYTYPE_ARRAY) {
-      conn = lupine_rpc_conn_for_deviceptr(copy.srcDevice);
-    } else {
-      conn = lupine_rpc_conn_for_current_context();
-    }
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy2D_v2) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    if (return_value == CUDA_SUCCESS) {
-      return_value = lupine_sync_mapped_device_to_host();
-    }
-    return return_value;
-  }
-  }
-  return CUDA_ERROR_INVALID_VALUE;
-}
-
-#ifdef cuMemcpy2D
-#undef cuMemcpy2D
-#endif
-extern "C" CUresult cuMemcpy2D(const CUDA_MEMCPY2D *pCopy) {
-  return cuMemcpy2D_v2(pCopy);
-}
-
-extern "C" CUresult cuMemcpy2DUnaligned_v2(const CUDA_MEMCPY2D *pCopy) {
-  if (pCopy == nullptr) {
-    return CUDA_ERROR_INVALID_VALUE;
-  }
-  CUDA_MEMCPY2D copy = *pCopy;
-  // UNIFIED addresses are client-space: a managed alias is already the
-  // server's address for that allocation, a plain host pointer's bytes must
-  // travel, and a device pointer is already server-valid.
-  if (copy.srcMemoryType == CU_MEMORYTYPE_UNIFIED) {
-    if (!lupine_is_managed_host_alias(copy.srcDevice) &&
-        lupine_copy_pointer_is_host(copy.srcDevice)) {
-      copy.srcHost = (const void *)copy.srcDevice;
-      copy.srcMemoryType = CU_MEMORYTYPE_HOST;
-    }
-  }
-  if (copy.dstMemoryType == CU_MEMORYTYPE_UNIFIED) {
-    if (!lupine_is_managed_host_alias(copy.dstDevice) &&
-        lupine_copy_pointer_is_host(copy.dstDevice)) {
-      copy.dstHost = (void *)copy.dstDevice;
-      copy.dstMemoryType = CU_MEMORYTYPE_HOST;
-    }
-  }
-  bool src_host = copy.srcMemoryType == CU_MEMORYTYPE_HOST;
-  bool dst_host = copy.dstMemoryType == CU_MEMORYTYPE_HOST;
-  lupine_copy_direction direction =
-      src_host ? (dst_host ? lupine_copy_direction::host_to_host
-                           : lupine_copy_direction::host_to_device)
-               : (dst_host ? lupine_copy_direction::device_to_host
-                           : lupine_copy_direction::device_to_device);
-  const char *src_base =
-      (const char *)copy.srcHost + copy.srcY * copy.srcPitch + copy.srcXInBytes;
-  char *dst_base =
-      (char *)copy.dstHost + copy.dstY * copy.dstPitch + copy.dstXInBytes;
-  CUresult return_value;
-  switch (direction) {
-  case lupine_copy_direction::host_to_host:
-    for (size_t row = 0; row < copy.Height; ++row) {
-      memmove(dst_base + row * copy.dstPitch, src_base + row * copy.srcPitch,
-              copy.WidthInBytes);
-    }
-    return CUDA_SUCCESS;
-  case lupine_copy_direction::host_to_device: {
-    conn_t *conn = copy.dstMemoryType == CU_MEMORYTYPE_ARRAY
-                       ? lupine_rpc_conn_for_current_context()
-                       : lupine_rpc_conn_for_deviceptr(copy.dstDevice);
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy2DUnaligned_v2) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_write_pitched(conn, src_base, copy.WidthInBytes, copy.Height,
-                          copy.srcPitch, 1, 0) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    if (return_value == CUDA_SUCCESS) {
-      return_value = lupine_sync_mapped_device_to_host();
-    }
-    return return_value;
-  }
-  case lupine_copy_direction::device_to_host: {
-    conn_t *conn = copy.srcMemoryType == CU_MEMORYTYPE_ARRAY
-                       ? lupine_rpc_conn_for_current_context()
-                       : lupine_rpc_conn_for_deviceptr(copy.srcDevice);
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy2DUnaligned_v2) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        (return_value == CUDA_SUCCESS &&
-         rpc_read_pitched(conn, dst_base, copy.WidthInBytes, copy.Height,
-                          copy.dstPitch, 1, 0) < 0) ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    return return_value;
-  }
-  case lupine_copy_direction::device_to_device: {
-    if (copy.srcMemoryType != CU_MEMORYTYPE_ARRAY &&
-        copy.dstMemoryType != CU_MEMORYTYPE_ARRAY &&
-        !lupine_deviceptrs_share_route(copy.dstDevice, copy.srcDevice)) {
-      // Different devices: the rows route through the client.
-      for (size_t row = 0; row < copy.Height; ++row) {
-        return_value = lupine_cuMemcpyDtoD_via_client(
-            copy.dstDevice + (copy.dstY + row) * copy.dstPitch +
-                copy.dstXInBytes,
-            copy.srcDevice + (copy.srcY + row) * copy.srcPitch +
-                copy.srcXInBytes,
-            copy.WidthInBytes, nullptr, false);
-        if (return_value != CUDA_SUCCESS) {
-          return return_value;
-        }
-      }
-      return lupine_sync_mapped_device_to_host();
-    }
-    conn_t *conn;
-    if (copy.dstMemoryType != CU_MEMORYTYPE_ARRAY) {
-      conn = lupine_rpc_conn_for_deviceptr(copy.dstDevice);
-    } else if (copy.srcMemoryType != CU_MEMORYTYPE_ARRAY) {
-      conn = lupine_rpc_conn_for_deviceptr(copy.srcDevice);
-    } else {
-      conn = lupine_rpc_conn_for_current_context();
-    }
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy2DUnaligned_v2) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    if (return_value == CUDA_SUCCESS) {
-      return_value = lupine_sync_mapped_device_to_host();
-    }
-    return return_value;
-  }
-  }
-  return CUDA_ERROR_INVALID_VALUE;
-}
-
-#ifdef cuMemcpy2DUnaligned
-#undef cuMemcpy2DUnaligned
-#endif
-extern "C" CUresult cuMemcpy2DUnaligned(const CUDA_MEMCPY2D *pCopy) {
-  return cuMemcpy2DUnaligned_v2(pCopy);
-}
-
-extern "C" CUresult cuMemcpy2DAsync_v2(const CUDA_MEMCPY2D *pCopy,
-                                       CUstream hStream) {
-  if (pCopy == nullptr) {
-    return CUDA_ERROR_INVALID_VALUE;
-  }
-  CUDA_MEMCPY2D copy = *pCopy;
-  // UNIFIED addresses are client-space: a managed alias is already the
-  // server's address for that allocation, a plain host pointer's bytes must
-  // travel, and a device pointer is already server-valid.
-  if (copy.srcMemoryType == CU_MEMORYTYPE_UNIFIED) {
-    if (!lupine_is_managed_host_alias(copy.srcDevice) &&
-        lupine_copy_pointer_is_host(copy.srcDevice)) {
-      copy.srcHost = (const void *)copy.srcDevice;
-      copy.srcMemoryType = CU_MEMORYTYPE_HOST;
-    }
-  }
-  if (copy.dstMemoryType == CU_MEMORYTYPE_UNIFIED) {
-    if (!lupine_is_managed_host_alias(copy.dstDevice) &&
-        lupine_copy_pointer_is_host(copy.dstDevice)) {
-      copy.dstHost = (void *)copy.dstDevice;
-      copy.dstMemoryType = CU_MEMORYTYPE_HOST;
-    }
-  }
-  bool src_host = copy.srcMemoryType == CU_MEMORYTYPE_HOST;
-  bool dst_host = copy.dstMemoryType == CU_MEMORYTYPE_HOST;
-  lupine_copy_direction direction =
-      src_host ? (dst_host ? lupine_copy_direction::host_to_host
-                           : lupine_copy_direction::host_to_device)
-               : (dst_host ? lupine_copy_direction::device_to_host
-                           : lupine_copy_direction::device_to_device);
-  const char *src_base =
-      (const char *)copy.srcHost + copy.srcY * copy.srcPitch + copy.srcXInBytes;
-  char *dst_base =
-      (char *)copy.dstHost + copy.dstY * copy.dstPitch + copy.dstXInBytes;
-  CUresult return_value;
-  switch (direction) {
-  case lupine_copy_direction::host_to_host:
-    for (size_t row = 0; row < copy.Height; ++row) {
-      memmove(dst_base + row * copy.dstPitch, src_base + row * copy.srcPitch,
-              copy.WidthInBytes);
-    }
-    return CUDA_SUCCESS;
-  case lupine_copy_direction::host_to_device: {
-    conn_t *conn = copy.dstMemoryType == CU_MEMORYTYPE_ARRAY
-                       ? lupine_rpc_conn_for_stream(hStream)
-                       : lupine_rpc_conn_for_deviceptr(copy.dstDevice);
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy2DAsync_v2) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_write_pitched(conn, src_base, copy.WidthInBytes, copy.Height,
-                          copy.srcPitch, 1, 0) < 0 ||
-        rpc_write(conn, &hStream, sizeof(hStream)) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    if (return_value == CUDA_SUCCESS) {
-      return_value = lupine_sync_mapped_device_to_host();
-    }
-    return return_value;
-  }
-  case lupine_copy_direction::device_to_host: {
-    conn_t *conn = copy.srcMemoryType == CU_MEMORYTYPE_ARRAY
-                       ? lupine_rpc_conn_for_stream(hStream)
-                       : lupine_rpc_conn_for_deviceptr(copy.srcDevice);
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy2DAsync_v2) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_write(conn, &hStream, sizeof(hStream)) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        (return_value == CUDA_SUCCESS &&
-         rpc_read_pitched(conn, dst_base, copy.WidthInBytes, copy.Height,
-                          copy.dstPitch, 1, 0) < 0) ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    return return_value;
-  }
-  case lupine_copy_direction::device_to_device: {
-    if (copy.srcMemoryType != CU_MEMORYTYPE_ARRAY &&
-        copy.dstMemoryType != CU_MEMORYTYPE_ARRAY &&
-        !lupine_deviceptrs_share_route(copy.dstDevice, copy.srcDevice)) {
-      // Different devices: the rows route through the client.
-      for (size_t row = 0; row < copy.Height; ++row) {
-        return_value = lupine_cuMemcpyDtoD_via_client(
-            copy.dstDevice + (copy.dstY + row) * copy.dstPitch +
-                copy.dstXInBytes,
-            copy.srcDevice + (copy.srcY + row) * copy.srcPitch +
-                copy.srcXInBytes,
-            copy.WidthInBytes, hStream, true);
-        if (return_value != CUDA_SUCCESS) {
-          return return_value;
-        }
-      }
-      return lupine_sync_mapped_device_to_host();
-    }
-    conn_t *conn;
-    if (copy.dstMemoryType != CU_MEMORYTYPE_ARRAY) {
-      conn = lupine_rpc_conn_for_deviceptr(copy.dstDevice);
-    } else if (copy.srcMemoryType != CU_MEMORYTYPE_ARRAY) {
-      conn = lupine_rpc_conn_for_deviceptr(copy.srcDevice);
-    } else {
-      conn = lupine_rpc_conn_for_stream(hStream);
-    }
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy2DAsync_v2) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_write(conn, &hStream, sizeof(hStream)) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    if (return_value == CUDA_SUCCESS) {
-      return_value = lupine_sync_mapped_device_to_host();
-    }
-    return return_value;
-  }
-  }
-  return CUDA_ERROR_INVALID_VALUE;
-}
-
-#ifdef cuMemcpy2DAsync
-#undef cuMemcpy2DAsync
-#endif
-extern "C" CUresult cuMemcpy2DAsync(const CUDA_MEMCPY2D *pCopy,
-                                    CUstream hStream) {
-  return cuMemcpy2DAsync_v2(pCopy, hStream);
-}
-
-#ifdef cuMemcpy2DAsync_ptsz
-#undef cuMemcpy2DAsync_ptsz
-#endif
-extern "C" CUresult cuMemcpy2DAsync_ptsz(const CUDA_MEMCPY2D *pCopy,
-                                         CUstream hStream) {
-  return cuMemcpy2DAsync_v2(pCopy, hStream);
-}
-
 static CUresult lupine_graph_mem_attribute_rpc(CUdevice device,
                                                CUgraphMem_attribute attr,
                                                cuuint64_t *value, bool set) {
@@ -5863,506 +5335,6 @@ extern "C" CUresult cuDeviceSetGraphMemAttribute(CUdevice device,
   cuuint64_t remote_value = 0;
   memcpy(&remote_value, value, sizeof(remote_value));
   return lupine_graph_mem_attribute_rpc(device, attr, &remote_value, true);
-}
-
-extern "C" CUresult cuMemcpy3D_v2(const CUDA_MEMCPY3D *pCopy) {
-  if (pCopy == nullptr) {
-    return CUDA_ERROR_INVALID_VALUE;
-  }
-  CUDA_MEMCPY3D copy = *pCopy;
-  // UNIFIED addresses are client-space: a managed alias is already the
-  // server's address for that allocation, a plain host pointer's bytes must
-  // travel, and a device pointer is already server-valid.
-  if (copy.srcMemoryType == CU_MEMORYTYPE_UNIFIED) {
-    if (!lupine_is_managed_host_alias(copy.srcDevice) &&
-        lupine_copy_pointer_is_host(copy.srcDevice)) {
-      copy.srcHost = (const void *)copy.srcDevice;
-      copy.srcMemoryType = CU_MEMORYTYPE_HOST;
-    }
-  }
-  if (copy.dstMemoryType == CU_MEMORYTYPE_UNIFIED) {
-    if (!lupine_is_managed_host_alias(copy.dstDevice) &&
-        lupine_copy_pointer_is_host(copy.dstDevice)) {
-      copy.dstHost = (void *)copy.dstDevice;
-      copy.dstMemoryType = CU_MEMORYTYPE_HOST;
-    }
-  }
-  bool src_host = copy.srcMemoryType == CU_MEMORYTYPE_HOST;
-  bool dst_host = copy.dstMemoryType == CU_MEMORYTYPE_HOST;
-  lupine_copy_direction direction =
-      src_host ? (dst_host ? lupine_copy_direction::host_to_host
-                           : lupine_copy_direction::host_to_device)
-               : (dst_host ? lupine_copy_direction::device_to_host
-                           : lupine_copy_direction::device_to_device);
-  size_t src_slice_pitch = copy.srcHeight * copy.srcPitch;
-  size_t dst_slice_pitch = copy.dstHeight * copy.dstPitch;
-  const char *src_base = (const char *)copy.srcHost +
-                         copy.srcZ * src_slice_pitch +
-                         copy.srcY * copy.srcPitch + copy.srcXInBytes;
-  char *dst_base = (char *)copy.dstHost + copy.dstZ * dst_slice_pitch +
-                   copy.dstY * copy.dstPitch + copy.dstXInBytes;
-  CUresult return_value;
-  switch (direction) {
-  case lupine_copy_direction::host_to_host:
-    for (size_t z = 0; z < copy.Depth; ++z) {
-      for (size_t row = 0; row < copy.Height; ++row) {
-        memmove(dst_base + z * dst_slice_pitch + row * copy.dstPitch,
-                src_base + z * src_slice_pitch + row * copy.srcPitch,
-                copy.WidthInBytes);
-      }
-    }
-    return CUDA_SUCCESS;
-  case lupine_copy_direction::host_to_device: {
-    conn_t *conn = copy.dstMemoryType == CU_MEMORYTYPE_ARRAY
-                       ? lupine_rpc_conn_for_current_context()
-                       : lupine_rpc_conn_for_deviceptr(copy.dstDevice);
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy3D_v2) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_write_pitched(conn, src_base, copy.WidthInBytes, copy.Height,
-                          copy.srcPitch, copy.Depth, src_slice_pitch) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    if (return_value == CUDA_SUCCESS) {
-      return_value = lupine_sync_mapped_device_to_host();
-    }
-    return return_value;
-  }
-  case lupine_copy_direction::device_to_host: {
-    conn_t *conn = copy.srcMemoryType == CU_MEMORYTYPE_ARRAY
-                       ? lupine_rpc_conn_for_current_context()
-                       : lupine_rpc_conn_for_deviceptr(copy.srcDevice);
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy3D_v2) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        (return_value == CUDA_SUCCESS &&
-         rpc_read_pitched(conn, dst_base, copy.WidthInBytes, copy.Height,
-                          copy.dstPitch, copy.Depth, dst_slice_pitch) < 0) ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    return return_value;
-  }
-  case lupine_copy_direction::device_to_device: {
-    if (copy.srcMemoryType != CU_MEMORYTYPE_ARRAY &&
-        copy.dstMemoryType != CU_MEMORYTYPE_ARRAY &&
-        !lupine_deviceptrs_share_route(copy.dstDevice, copy.srcDevice)) {
-      // Different devices: the rows route through the client.
-      for (size_t z = 0; z < copy.Depth; ++z) {
-        for (size_t row = 0; row < copy.Height; ++row) {
-          return_value = lupine_cuMemcpyDtoD_via_client(
-              copy.dstDevice + (copy.dstZ + z) * dst_slice_pitch +
-                  (copy.dstY + row) * copy.dstPitch + copy.dstXInBytes,
-              copy.srcDevice + (copy.srcZ + z) * src_slice_pitch +
-                  (copy.srcY + row) * copy.srcPitch + copy.srcXInBytes,
-              copy.WidthInBytes, nullptr, false);
-          if (return_value != CUDA_SUCCESS) {
-            return return_value;
-          }
-        }
-      }
-      return lupine_sync_mapped_device_to_host();
-    }
-    conn_t *conn;
-    if (copy.dstMemoryType != CU_MEMORYTYPE_ARRAY) {
-      conn = lupine_rpc_conn_for_deviceptr(copy.dstDevice);
-    } else if (copy.srcMemoryType != CU_MEMORYTYPE_ARRAY) {
-      conn = lupine_rpc_conn_for_deviceptr(copy.srcDevice);
-    } else {
-      conn = lupine_rpc_conn_for_current_context();
-    }
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy3D_v2) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    if (return_value == CUDA_SUCCESS) {
-      return_value = lupine_sync_mapped_device_to_host();
-    }
-    return return_value;
-  }
-  }
-  return CUDA_ERROR_INVALID_VALUE;
-}
-
-
-extern "C" CUresult cuMemcpy3DAsync_v2(const CUDA_MEMCPY3D *pCopy,
-                                       CUstream hStream) {
-  if (pCopy == nullptr) {
-    return CUDA_ERROR_INVALID_VALUE;
-  }
-  CUDA_MEMCPY3D copy = *pCopy;
-  // UNIFIED addresses are client-space: a managed alias is already the
-  // server's address for that allocation, a plain host pointer's bytes must
-  // travel, and a device pointer is already server-valid.
-  if (copy.srcMemoryType == CU_MEMORYTYPE_UNIFIED) {
-    if (!lupine_is_managed_host_alias(copy.srcDevice) &&
-        lupine_copy_pointer_is_host(copy.srcDevice)) {
-      copy.srcHost = (const void *)copy.srcDevice;
-      copy.srcMemoryType = CU_MEMORYTYPE_HOST;
-    }
-  }
-  if (copy.dstMemoryType == CU_MEMORYTYPE_UNIFIED) {
-    if (!lupine_is_managed_host_alias(copy.dstDevice) &&
-        lupine_copy_pointer_is_host(copy.dstDevice)) {
-      copy.dstHost = (void *)copy.dstDevice;
-      copy.dstMemoryType = CU_MEMORYTYPE_HOST;
-    }
-  }
-  bool src_host = copy.srcMemoryType == CU_MEMORYTYPE_HOST;
-  bool dst_host = copy.dstMemoryType == CU_MEMORYTYPE_HOST;
-  lupine_copy_direction direction =
-      src_host ? (dst_host ? lupine_copy_direction::host_to_host
-                           : lupine_copy_direction::host_to_device)
-               : (dst_host ? lupine_copy_direction::device_to_host
-                           : lupine_copy_direction::device_to_device);
-  size_t src_slice_pitch = copy.srcHeight * copy.srcPitch;
-  size_t dst_slice_pitch = copy.dstHeight * copy.dstPitch;
-  const char *src_base =
-      (const char *)copy.srcHost + copy.srcZ * src_slice_pitch +
-      copy.srcY * copy.srcPitch + copy.srcXInBytes;
-  char *dst_base = (char *)copy.dstHost + copy.dstZ * dst_slice_pitch +
-                   copy.dstY * copy.dstPitch + copy.dstXInBytes;
-  CUresult return_value;
-  switch (direction) {
-  case lupine_copy_direction::host_to_host:
-    for (size_t z = 0; z < copy.Depth; ++z) {
-      for (size_t row = 0; row < copy.Height; ++row) {
-        memmove(dst_base + z * dst_slice_pitch + row * copy.dstPitch,
-                src_base + z * src_slice_pitch + row * copy.srcPitch,
-                copy.WidthInBytes);
-      }
-    }
-    return CUDA_SUCCESS;
-  case lupine_copy_direction::host_to_device: {
-    conn_t *conn = copy.dstMemoryType == CU_MEMORYTYPE_ARRAY
-                       ? lupine_rpc_conn_for_stream(hStream)
-                       : lupine_rpc_conn_for_deviceptr(copy.dstDevice);
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy3DAsync_v2) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_write_pitched(conn, src_base, copy.WidthInBytes, copy.Height,
-                          copy.srcPitch, copy.Depth, src_slice_pitch) < 0 ||
-        rpc_write(conn, &hStream, sizeof(hStream)) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    if (return_value == CUDA_SUCCESS) {
-      return_value = lupine_sync_mapped_device_to_host();
-    }
-    return return_value;
-  }
-  case lupine_copy_direction::device_to_host: {
-    conn_t *conn = copy.srcMemoryType == CU_MEMORYTYPE_ARRAY
-                       ? lupine_rpc_conn_for_stream(hStream)
-                       : lupine_rpc_conn_for_deviceptr(copy.srcDevice);
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy3DAsync_v2) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_write(conn, &hStream, sizeof(hStream)) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        (return_value == CUDA_SUCCESS &&
-         rpc_read_pitched(conn, dst_base, copy.WidthInBytes, copy.Height,
-                          copy.dstPitch, copy.Depth, dst_slice_pitch) < 0) ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    return return_value;
-  }
-  case lupine_copy_direction::device_to_device: {
-    if (copy.srcMemoryType != CU_MEMORYTYPE_ARRAY &&
-        copy.dstMemoryType != CU_MEMORYTYPE_ARRAY &&
-        !lupine_deviceptrs_share_route(copy.dstDevice, copy.srcDevice)) {
-      // Different devices: the rows route through the client.
-      for (size_t z = 0; z < copy.Depth; ++z) {
-        for (size_t row = 0; row < copy.Height; ++row) {
-          return_value = lupine_cuMemcpyDtoD_via_client(
-              copy.dstDevice + (copy.dstZ + z) * dst_slice_pitch +
-                  (copy.dstY + row) * copy.dstPitch + copy.dstXInBytes,
-              copy.srcDevice + (copy.srcZ + z) * src_slice_pitch +
-                  (copy.srcY + row) * copy.srcPitch + copy.srcXInBytes,
-              copy.WidthInBytes, hStream, true);
-          if (return_value != CUDA_SUCCESS) {
-            return return_value;
-          }
-        }
-      }
-      return lupine_sync_mapped_device_to_host();
-    }
-    conn_t *conn;
-    if (copy.dstMemoryType != CU_MEMORYTYPE_ARRAY) {
-      conn = lupine_rpc_conn_for_deviceptr(copy.dstDevice);
-    } else if (copy.srcMemoryType != CU_MEMORYTYPE_ARRAY) {
-      conn = lupine_rpc_conn_for_deviceptr(copy.srcDevice);
-    } else {
-      conn = lupine_rpc_conn_for_stream(hStream);
-    }
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy3DAsync_v2) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_write(conn, &hStream, sizeof(hStream)) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    if (return_value == CUDA_SUCCESS) {
-      return_value = lupine_sync_mapped_device_to_host();
-    }
-    return return_value;
-  }
-  }
-  return CUDA_ERROR_INVALID_VALUE;
-}
-
-
-extern "C" CUresult cuMemcpy3DPeer(const CUDA_MEMCPY3D_PEER *pCopy) {
-  if (pCopy == nullptr) {
-    return CUDA_ERROR_INVALID_VALUE;
-  }
-  // Both context handles must live on one connection; the copy then behaves
-  // like an ordinary one there. Peers on different connections are not
-  // supported.
-  lupine_route src_route = lupine_route_for_context(pCopy->srcContext);
-  lupine_route dst_route = lupine_route_for_context(pCopy->dstContext);
-  if (lupine_route_identity(src_route) != lupine_route_identity(dst_route) ||
-      lupine_route_is_local(dst_route)) {
-    return CUDA_ERROR_NOT_SUPPORTED;
-  }
-  conn_t *conn = lupine_route_remote_conn(dst_route);
-  CUDA_MEMCPY3D_PEER copy = *pCopy;
-  // UNIFIED addresses are client-space: a managed alias is already the
-  // server's address for that allocation, a plain host pointer's bytes must
-  // travel, and a device pointer is already server-valid.
-  if (copy.srcMemoryType == CU_MEMORYTYPE_UNIFIED) {
-    if (!lupine_is_managed_host_alias(copy.srcDevice) &&
-        lupine_copy_pointer_is_host(copy.srcDevice)) {
-      copy.srcHost = (const void *)copy.srcDevice;
-      copy.srcMemoryType = CU_MEMORYTYPE_HOST;
-    }
-  }
-  if (copy.dstMemoryType == CU_MEMORYTYPE_UNIFIED) {
-    if (!lupine_is_managed_host_alias(copy.dstDevice) &&
-        lupine_copy_pointer_is_host(copy.dstDevice)) {
-      copy.dstHost = (void *)copy.dstDevice;
-      copy.dstMemoryType = CU_MEMORYTYPE_HOST;
-    }
-  }
-  bool src_host = copy.srcMemoryType == CU_MEMORYTYPE_HOST;
-  bool dst_host = copy.dstMemoryType == CU_MEMORYTYPE_HOST;
-  lupine_copy_direction direction =
-      src_host ? (dst_host ? lupine_copy_direction::host_to_host
-                           : lupine_copy_direction::host_to_device)
-               : (dst_host ? lupine_copy_direction::device_to_host
-                           : lupine_copy_direction::device_to_device);
-  size_t src_slice_pitch = copy.srcHeight * copy.srcPitch;
-  size_t dst_slice_pitch = copy.dstHeight * copy.dstPitch;
-  const char *src_base =
-      (const char *)copy.srcHost + copy.srcZ * src_slice_pitch +
-      copy.srcY * copy.srcPitch + copy.srcXInBytes;
-  char *dst_base = (char *)copy.dstHost + copy.dstZ * dst_slice_pitch +
-                   copy.dstY * copy.dstPitch + copy.dstXInBytes;
-  CUresult return_value;
-  switch (direction) {
-  case lupine_copy_direction::host_to_host:
-    for (size_t z = 0; z < copy.Depth; ++z) {
-      for (size_t row = 0; row < copy.Height; ++row) {
-        memmove(dst_base + z * dst_slice_pitch + row * copy.dstPitch,
-                src_base + z * src_slice_pitch + row * copy.srcPitch,
-                copy.WidthInBytes);
-      }
-    }
-    return CUDA_SUCCESS;
-  case lupine_copy_direction::host_to_device: {
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy3DPeer) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_write_pitched(conn, src_base, copy.WidthInBytes, copy.Height,
-                          copy.srcPitch, copy.Depth, src_slice_pitch) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    if (return_value == CUDA_SUCCESS) {
-      return_value = lupine_sync_mapped_device_to_host();
-    }
-    return return_value;
-  }
-  case lupine_copy_direction::device_to_host: {
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy3DPeer) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        (return_value == CUDA_SUCCESS &&
-         rpc_read_pitched(conn, dst_base, copy.WidthInBytes, copy.Height,
-                          copy.dstPitch, copy.Depth, dst_slice_pitch) < 0) ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    return return_value;
-  }
-  case lupine_copy_direction::device_to_device: {
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy3DPeer) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    if (return_value == CUDA_SUCCESS) {
-      return_value = lupine_sync_mapped_device_to_host();
-    }
-    return return_value;
-  }
-  }
-  return CUDA_ERROR_INVALID_VALUE;
-}
-
-
-extern "C" CUresult cuMemcpy3DPeerAsync(const CUDA_MEMCPY3D_PEER *pCopy,
-                                        CUstream hStream) {
-  if (pCopy == nullptr) {
-    return CUDA_ERROR_INVALID_VALUE;
-  }
-  // Both context handles must live on one connection; the copy then behaves
-  // like an ordinary one there. Peers on different connections are not
-  // supported.
-  lupine_route src_route = lupine_route_for_context(pCopy->srcContext);
-  lupine_route dst_route = lupine_route_for_context(pCopy->dstContext);
-  if (lupine_route_identity(src_route) != lupine_route_identity(dst_route) ||
-      lupine_route_is_local(dst_route)) {
-    return CUDA_ERROR_NOT_SUPPORTED;
-  }
-  conn_t *conn = lupine_route_remote_conn(dst_route);
-  CUDA_MEMCPY3D_PEER copy = *pCopy;
-  // UNIFIED addresses are client-space: a managed alias is already the
-  // server's address for that allocation, a plain host pointer's bytes must
-  // travel, and a device pointer is already server-valid.
-  if (copy.srcMemoryType == CU_MEMORYTYPE_UNIFIED) {
-    if (!lupine_is_managed_host_alias(copy.srcDevice) &&
-        lupine_copy_pointer_is_host(copy.srcDevice)) {
-      copy.srcHost = (const void *)copy.srcDevice;
-      copy.srcMemoryType = CU_MEMORYTYPE_HOST;
-    }
-  }
-  if (copy.dstMemoryType == CU_MEMORYTYPE_UNIFIED) {
-    if (!lupine_is_managed_host_alias(copy.dstDevice) &&
-        lupine_copy_pointer_is_host(copy.dstDevice)) {
-      copy.dstHost = (void *)copy.dstDevice;
-      copy.dstMemoryType = CU_MEMORYTYPE_HOST;
-    }
-  }
-  bool src_host = copy.srcMemoryType == CU_MEMORYTYPE_HOST;
-  bool dst_host = copy.dstMemoryType == CU_MEMORYTYPE_HOST;
-  lupine_copy_direction direction =
-      src_host ? (dst_host ? lupine_copy_direction::host_to_host
-                           : lupine_copy_direction::host_to_device)
-               : (dst_host ? lupine_copy_direction::device_to_host
-                           : lupine_copy_direction::device_to_device);
-  size_t src_slice_pitch = copy.srcHeight * copy.srcPitch;
-  size_t dst_slice_pitch = copy.dstHeight * copy.dstPitch;
-  const char *src_base =
-      (const char *)copy.srcHost + copy.srcZ * src_slice_pitch +
-      copy.srcY * copy.srcPitch + copy.srcXInBytes;
-  char *dst_base = (char *)copy.dstHost + copy.dstZ * dst_slice_pitch +
-                   copy.dstY * copy.dstPitch + copy.dstXInBytes;
-  CUresult return_value;
-  switch (direction) {
-  case lupine_copy_direction::host_to_host:
-    for (size_t z = 0; z < copy.Depth; ++z) {
-      for (size_t row = 0; row < copy.Height; ++row) {
-        memmove(dst_base + z * dst_slice_pitch + row * copy.dstPitch,
-                src_base + z * src_slice_pitch + row * copy.srcPitch,
-                copy.WidthInBytes);
-      }
-    }
-    return CUDA_SUCCESS;
-  case lupine_copy_direction::host_to_device: {
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy3DPeerAsync) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_write_pitched(conn, src_base, copy.WidthInBytes, copy.Height,
-                          copy.srcPitch, copy.Depth, src_slice_pitch) < 0 ||
-        rpc_write(conn, &hStream, sizeof(hStream)) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    if (return_value == CUDA_SUCCESS) {
-      return_value = lupine_sync_mapped_device_to_host();
-    }
-    return return_value;
-  }
-  case lupine_copy_direction::device_to_host: {
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy3DPeerAsync) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_write(conn, &hStream, sizeof(hStream)) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        (return_value == CUDA_SUCCESS &&
-         rpc_read_pitched(conn, dst_base, copy.WidthInBytes, copy.Height,
-                          copy.dstPitch, copy.Depth, dst_slice_pitch) < 0) ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    return return_value;
-  }
-  case lupine_copy_direction::device_to_device: {
-    if (lupine_prepare_rpc(conn) < 0 ||
-        rpc_write_start_request(conn, RPC_cuMemcpy3DPeerAsync) < 0 ||
-        rpc_write(conn, &copy, sizeof(copy)) < 0 ||
-        rpc_write(conn, &hStream, sizeof(hStream)) < 0 ||
-        rpc_wait_for_response(conn) < 0 ||
-        rpc_read(conn, &return_value, sizeof(return_value)) < 0 ||
-        rpc_read_end(conn) < 0) {
-      return CUDA_ERROR_DEVICE_UNAVAILABLE;
-    }
-    if (return_value == CUDA_SUCCESS) {
-      return_value = lupine_sync_mapped_device_to_host();
-    }
-    return return_value;
-  }
-  }
-  return CUDA_ERROR_INVALID_VALUE;
-}
-
-#ifdef cuMemcpy3D
-#undef cuMemcpy3D
-#endif
-extern "C" CUresult cuMemcpy3D(const CUDA_MEMCPY3D *pCopy) {
-  return cuMemcpy3D_v2(pCopy);
-}
-
-#ifdef cuMemcpy3DAsync
-#undef cuMemcpy3DAsync
-#endif
-extern "C" CUresult cuMemcpy3DAsync(const CUDA_MEMCPY3D *pCopy,
-                                    CUstream hStream) {
-  return cuMemcpy3DAsync_v2(pCopy, hStream);
-}
-
-#ifdef cuMemcpy3DAsync_ptsz
-#undef cuMemcpy3DAsync_ptsz
-#endif
-extern "C" CUresult cuMemcpy3DAsync_ptsz(const CUDA_MEMCPY3D *pCopy,
-                                         CUstream hStream) {
-  return cuMemcpy3DAsync_v2(pCopy, hStream);
 }
 
 static CUresult
