+++ cuda_client_memcpy.h
